pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "AlpacaTrading"
version = "2.0.0"
requires-python = ">=3.13"
dependencies = [
    "alpaca-py>=0.43.2",
    "contourpy==1.3.3",
    "cycler==0.12.1",
    "fonttools==4.60.1",
    "ipykernel>=6.30.1",
    "kiwisolver==1.4.9",
    "matplotlib==3.10.7",
    "memory-profiler>=0.61.0",
    "numpy==2.3.3",
    "packaging==25.0",
    "pandas>=2.3.3",
    "pillow==11.3.0",
    "pyparsing==3.2.5",
    "pytest>=9.0.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv>=1.2.1",
    "seaborn>=0.13.2",
    "six==1.17.0",
    "yfinance>=0.2.66",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
AlpacaTrading = ["py.typed"]

[tool.pytest.ini_options]
pythonpath = ["src"]
